Rename HeapRegionStateTable to CopyForwardRegionLookupTable
- Introduce new getter and setter functions to ease the OpenJ9
  transition.
