+++ fvtest/gctest/TestCopyForwardRegionLookupTable.cpp
@@ -19,7 +19,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0 OR GPL-2.0 WITH Classpath-exception-2.0 OR LicenseRef-GPL-2.0 WITH Assembly-exception
  *******************************************************************************/
 
-#include "HeapRegionStateTable.hpp"
+#include "CopyForwardRegionLookupTable.hpp"
 #include "gcTestHelpers.hpp"
 
 #include <Forge.hpp>
@@ -28,7 +28,7 @@
 
 using namespace OMR::GC;
 
-TEST(TestHeapRegionStateTable, HeapRegionStateTable)
+TEST(TestCopyForwardRegionLookupTable, CopyForwardRegionLookupTable)
 {
     uintptr_t heapBase = 0x100;
     uintptr_t regionShift = 1;
@@ -37,7 +37,7 @@ TEST(TestHeapRegionStateTable, HeapRegionStateTable)
     Forge forge;
     ASSERT_TRUE(forge.initialize(gcTestEnv->getPortLibrary()));
 
-    HeapRegionStateTable table;
+    CopyForwardRegionLookupTable table;
     ASSERT_TRUE(table.initialize(&forge, heapBase, regionShift, regionCount));
 
     EXPECT_EQ(table.getIndex((void *)0x100), 0);
